find logs

+++ src/main/java/com/github/leandrolimasi/dto/AppArgumentsDTO.java
@@ -24,7 +24,7 @@ public class AppArgumentsDTO {
 
     private Duration duration;
 
-    private Integer threshold;
+    private Long threshold;
 
     private File logFile;
